feat: Add DeleteApplicationResources
This method will let clients to delete an application and its associated resources even in the case the application wasn’t successfully created in the first place.

Useful for `meroxa apps remove app-name`, and for deploying over an existing application (we’d be tearing down existing resources before creating new ones with the same name).